[feat/#59]: 로그인 후 라우팅 수정
