+++ components/study/Timer.tsx
@@ -11,6 +11,9 @@ import { useSubjectStore } from '@/store/subjectStore';
 import Cookies from 'js-cookie';
 import { ErrorResponse } from '@/types/commonType';
 import { postTimer } from '@/api/studyApi';
+import { useUserInfoStore } from '@/store/memberStore';
+import useSWR from 'swr';
+import { getUserInfo } from '@/api/memberApi';
 
 interface ITimer {
   maxTime: number;
@@ -27,6 +30,16 @@ export default function Timer({ maxTime, currentTime }: ITimer) {
   const [time, setTime] = useState(currentTime); // 초 단위
   const [progress, setProgress] = useState((currentTime / maxTime) * 100);
   const [remainingTime, setRemainingTime] = useState(maxTime - (currentTime % maxTime));
+
+  // 유저정보조회
+  const accessToken = Cookies.get('access_token');
+  const { setUserInfo } = useUserInfoStore();
+  const { data } = useSWR(accessToken ? ['userInfo'] : null, async () => {
+    const result = await getUserInfo();
+    const storedUserInfo = { name: result.name, course: result.course, image: result.image };
+    localStorage.setItem('userInfo', JSON.stringify(storedUserInfo));
+    setUserInfo(storedUserInfo);
+  });
 
   const startTimeRef = useRef<number | null>(null);
   const animationFrameRef = useRef<number | null>(null);
